Splitted script into lib and cli
