+++ .gitignore
@@ -15,6 +15,5 @@ docker-compose.yml
 # Generated stylesheet
 static/style.css
 
-# Files generated by tests
-tests/compare_output/
-tests/current_output/
+# Temporary files
+tmp
